Move pubsub-client to new crate
