+++ client/src/lib.rs
@@ -16,6 +16,7 @@ pub mod rpc_client;
 pub mod rpc_config;
 pub mod rpc_custom_error;
 pub mod rpc_deprecated_config;
+pub mod rpc_error_object;
 pub mod rpc_filter;
 pub mod rpc_request;
 pub mod rpc_response;
