+++ client/src/rpc_error_object.rs
@@ -1,0 +1,5 @@
+#[derive(Deserialize, Debug)]
+pub struct RpcErrorObject {
+    pub code: i64,
+    pub message: String,
+}
