+++ pubsub-client/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "solana-pubsub-client"
+version = "1.12.0"
+description = "Solana Pubsub Client"
+authors = ["Solana Maintainers <maintainers@solana.foundation>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+documentation = "https://docs.rs/solana-pubsub-client"
+edition = "2021"
+
+[dependencies]
+crossbeam-channel = "0.5"
+futures-util = "0.3.21"
+log = "0.4.17"
+reqwest = { version = "0.11.11", default-features = false, features = ["blocking", "brotli", "deflate", "gzip", "rustls-tls", "json"] }
+semver = "1.0.13"
+serde = "1.0.143"
+serde_derive = "1.0.103"
+serde_json = "1.0.83"
+solana-account-decoder = { path = "../account-decoder", version = "=1.12.0" }
+solana-client = { path = "../client", version = "=1.12.0" }
+solana-sdk = { path = "../sdk", version = "=1.12.0" }
+thiserror = "1.0"
+tokio = { version = "1", features = ["full"] }
+tokio-stream = "0.1.9"
+tokio-tungstenite = { version = "0.17.2", features = ["rustls-tls-webpki-roots"] }
+tungstenite = { version = "0.17.2", features = ["rustls-tls-webpki-roots"] }
+url = "2.2.2"
+
+[dev-dependencies]
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
