+++ pubsub-client/src/lib.rs
@@ -1,0 +1,4 @@
+#![allow(clippy::integer_arithmetic)]
+
+pub mod nonblocking;
+pub mod pubsub_client;
