+++ tpu-client/Cargo.toml
@@ -30,6 +30,7 @@ solana-client = { path = "../client", version = "=1.12.0" }
 solana-measure = { path = "../measure", version = "=1.12.0" }
 solana-metrics = { path = "../metrics", version = "=1.12.0" }
 solana-net-utils = { path = "../net-utils", version = "=1.12.0" }
+solana-pubsub-client = { path = "../pubsub-client", version = "=1.12.0" }
 solana-sdk = { path = "../sdk", version = "=1.12.0" }
 solana-streamer = { path = "../streamer", version = "=1.12.0" }
 thiserror = "1.0"
